requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   groq
    #   httpx
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.3
    # via job-email-classifier (pyproject.toml)
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
cryptography==46.0.4
    # via google-auth
distro==1.9.0
    # via groq
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-api-core==2.29.0
    # via google-api-python-client
google-api-python-client==2.188.0
    # via job-email-classifier (pyproject.toml)
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-auth-oauthlib==1.2.4
    # via job-email-classifier (pyproject.toml)
googleapis-common-protos==1.72.0
    # via google-api-core
groq==1.0.0
    # via job-email-classifier (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via groq
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.15.0
    # via altair
numpy==2.4.1
    # via
    #   pandas
    #   pydeck
    #   streamlit
oauthlib==3.3.1
    # via requests-oauthlib
packaging==26.0
    # via
    #   altair
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.1.0
    # via streamlit
proto-plus==1.27.0
    # via google-api-core
protobuf==6.33.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   streamlit
pyarrow==23.0.0
    # via streamlit
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via groq
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyparsing==3.3.2
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via job-email-classifier (pyproject.toml)
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   google-api-core
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via groq
soupsieve==2.8.3
    # via beautifulsoup4
streamlit==1.53.1
    # via job-email-classifier (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   beautifulsoup4
    #   groq
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
